Update README.md
When you organize your data files from the EZ_Tracker software, it is often useful to compile all of the CSV files into a single Excel file for ease of statistical analysis. The sheet names in the Excel file can be used to identify the source of the data. This code helps you do that by compiling a list of CSV files in a directory and its subdirectories into a single Excel file, using the names of the folders containing the files as the sheet names.